requirements.txt
google-genai
youtube-transcript-api
dotenv
google-adk
